Final base of exercise

+++ src/main/java/com/sngular/devTraining2021/exercises/AlgorithmName.java
@@ -1,7 +1,7 @@
 package com.sngular.devTrining2021.exercises;
 
 public class AlgorithmName {
-    public static long orderNumber(long number){
+    public static int orderNumber(int number){
         //Implements the function with the algorithm
         return 0;
     }
